Generate a user key when missing and send it to keyserver
When the keyserver doesn't have a user key for a particular user,
BrowserID will generate a random one and then send it to the
keyserver for storage.

Note that right now we are sending plain keys back and forth.
Eventually, these plain keys will be replaced by a user key that
is wrapped by a password-derived key.