Pull the dialog background gradient into a separate div
This allows the background image to be in a single property, on
the body directly. It is a pre-requisite for inlining the image
into the stylesheet.

Warning: it is VISUALLY DIFFERENT from the original (take a close
look at the footer). I'm not sure why, but it should definitely
NOT be merged anywhere yet.