Prefetch dialog components from the communication_iframe
When the browser is idle, it will start fetching and caching the
dialog's Javascript, CSS and images. This works on IE9, Firefox
and Chromium.

The trick is to make sure that the communication iframe is open
while the user is busy reading the RP's site. When using the watch
API, this happens automatically at the end of the .watch() call.