client-sessions: stop specifying maxAge and allow ephemeral cookies
The latest version of client-sessions automatically sets the
HTTP expiry on the cookie to match the duration of the signed
payload. Therefore it's no longer necessar to set cookie.maxAge.

The ephemeral parameter to setDuration() is used to change the
cookie to an ephemeral session cookie from a regular cookie and
vice-versa.